+++ lib/src/eval/compiler/macros/branch.dart
@@ -1,5 +1,6 @@
 import 'package:dart_eval/src/eval/compiler/context.dart';
 import 'package:dart_eval/src/eval/compiler/macros/macro.dart';
+import 'package:dart_eval/src/eval/compiler/model/label.dart';
 import 'package:dart_eval/src/eval/compiler/statement/statement.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
 import 'package:dart_eval/src/eval/runtime/runtime.dart';
@@ -21,7 +22,17 @@ StatementInfo macroBranch(
 
   ctx.inferTypes();
   ctx.beginAllocScope();
+  final label = CompilerLabel(LabelType.branch, -1, (_ctx) {
+    _ctx.endAllocScopeQuiet();
+    if (!resolveStateToThen) {
+      _ctx.resolveBranchStateDiscontinuity(_initialState);
+    }
+    _ctx.endAllocScopeQuiet();
+    return -1;
+  });
+  ctx.labels.add(label);
   final thenResult = thenBranch(ctx, expectedReturnType);
+  ctx.labels.removeLast();
   ctx.endAllocScope();
   ctx.uninferTypes();
 
@@ -41,7 +52,15 @@ StatementInfo macroBranch(
 
   if (elseBranch != null) {
     ctx.beginAllocScope();
+    final label = CompilerLabel(LabelType.branch, -1, (_ctx) {
+      ctx.endAllocScope();
+      ctx.resolveBranchStateDiscontinuity(_initialState);
+      ctx.endAllocScope();
+      return -1;
+    });
+    ctx.labels.add(label);
     final elseResult = elseBranch(ctx, expectedReturnType);
+    ctx.labels.removeLast();
     ctx.endAllocScope();
     ctx.resolveBranchStateDiscontinuity(_initialState);
     ctx.rewriteOp(rewriteOut!, JumpConstant.make(ctx.out.length), 0);