+++ lib/src/eval/compiler/statement/statement.dart
@@ -4,6 +4,7 @@ import 'package:dart_eval/src/eval/compiler/context.dart';
 import 'package:dart_eval/src/eval/compiler/errors.dart';
 import 'package:dart_eval/src/eval/compiler/expression/expression.dart';
 import 'package:dart_eval/src/eval/compiler/statement/assert.dart';
+import 'package:dart_eval/src/eval/compiler/statement/break.dart';
 import 'package:dart_eval/src/eval/compiler/statement/do.dart';
 import 'package:dart_eval/src/eval/compiler/statement/for.dart';
 import 'package:dart_eval/src/eval/compiler/statement/if.dart';
@@ -41,6 +42,8 @@ StatementInfo compileStatement(
     return compileTryStatement(s, ctx, expectedReturnType);
   } else if (s is AssertStatement) {
     return compileAssertStatement(s, ctx, expectedReturnType);
+  } else if (s is BreakStatement) {
+    return compileBreakStatement(s, ctx);
   } else {
     throw CompileError('Unknown statement type ${s.runtimeType}');
   }